feat(install): improve platform compatibility
First, enable action to work correctly on all supported platforms.
Check the OS and Architecture of the runner, download the correct version
if available, and bail out with an appropriate error if not supported.

Second, behave like a github setup- tool.
This means download to a versioned subdirectory in the GHA tool cache, and
don't download again if the binary is already there.
(also delete the tarball as part of the install).
Allow the user to optionally select the cog version